Revert "Merge remote-tracking branch 'upstream/develop' into Shuffle-Roll"
This reverts commit f56452d10d35fed3a5968fc8b558a1955aeaadad, reversing
changes made to 447272879a099437a18424e359f0c673cb88d15d.

+++ soh/soh/frame_interpolation.cpp
@@ -1,3 +1,5 @@
+#include <libultraship/bridge.h>
+
 #include <vector>
 #include <map>
 #include <unordered_map>
@@ -235,19 +237,19 @@ struct InterpolateCtx {
     float interpolate_angle(f32 o, f32 n) {
         if (o == n)
             return n;
-        o = fmodf(o, static_cast<f32>(2.0f * M_PI));
+        o = fmodf(o, 2 * M_PI);
         if (o < 0.0f) {
-            o += static_cast<f32>(2.0f * M_PI);
+            o += 2 * M_PI;
         }
-        n = fmodf(n, static_cast<f32>(2.0f * M_PI));
+        n = fmodf(n, 2 * M_PI);
         if (n < 0.0f) {
-            n += static_cast<f32>(2.0f * M_PI);
+            n += 2 * M_PI;
         }
         if (fabsf(o - n) > M_PI) {
             if (o < n) {
-                o += static_cast<f32>(2.0f * M_PI);
+                o += 2 * M_PI;
             } else {
-                n += static_cast<f32>(2.0f * M_PI);
+                n += 2 * M_PI;
             }
         }
         if (fabsf(o - n) > M_PI / 2) {
@@ -727,7 +729,7 @@ static bool invert_matrix(const float m[16], float invOut[16]) {
         return false;
     }
 
-    det = 1.0f / det;
+    det = 1.0 / det;
 
     for (i = 0; i < 16; i++) {
         invOut[i] = inv[i] * det;
