+++ soh/soh/resource/importer/PathFactory.cpp
@@ -3,7 +3,6 @@
 #include "soh/resource/logging/PathLogger.h"
 #include "spdlog/spdlog.h"
 #include <tinyxml2.h>
-#include <libultraship/bridge/consolevariablebridge.h>
 
 namespace SOH {
 std::shared_ptr<Ship::IResource>
@@ -84,7 +83,7 @@ ResourceFactoryXMLPathV0::ReadResource(std::shared_ptr<Ship::File> file,
 
         PathData pathDataEntry;
         // pathDataEntry.count = pointCount;
-        pathDataEntry.count = static_cast<u8>(points.size());
+        pathDataEntry.count = points.size();
 
         path->paths.push_back(points);
         pathDataEntry.points = path->paths.back().data();
@@ -94,7 +93,7 @@ ResourceFactoryXMLPathV0::ReadResource(std::shared_ptr<Ship::File> file,
         pathDataElement = pathDataElement->NextSiblingElement();
     }
 
-    path->numPaths = static_cast<u32>(path->paths.size());
+    path->numPaths = path->paths.size();
 
     return path;
 };
