+++ soh/soh/resource/importer/scenecommand/SetObjectListFactory.cpp
@@ -41,7 +41,7 @@ std::shared_ptr<Ship::IResource> SetObjectListFactoryXML::ReadResource(std::shar
         child = child->NextSiblingElement();
     }
 
-    setObjectList->numObjects = static_cast<u32>(setObjectList->objects.size());
+    setObjectList->numObjects = setObjectList->objects.size();
 
     return setObjectList;
 }
