+++ soh/soh/resource/importer/scenecommand/SetRoomListFactory.cpp
@@ -1,8 +1,7 @@
 #include "soh/resource/importer/scenecommand/SetRoomListFactory.h"
 #include "soh/resource/type/scenecommand/SetRoomList.h"
 #include "soh/resource/logging/SceneCommandLoggers.h"
-#include <libultraship/libultra/types.h>
-#include <spdlog/spdlog.h>
+#include "spdlog/spdlog.h"
 #include <tinyxml2.h>
 
 namespace SOH {
@@ -58,7 +57,7 @@ std::shared_ptr<Ship::IResource> SetRoomListFactoryXML::ReadResource(std::shared
         child = child->NextSiblingElement();
     }
 
-    setRoomList->numRooms = static_cast<u32>(setRoomList->rooms.size());
+    setRoomList->numRooms = setRoomList->rooms.size();
 
     return setRoomList;
 }
